refactor mavlink_param_value out of mavlink_parameters - I think this improves readability of the mavlink_parameters class (even after refactoring that out the mavlink_parameters.cpp file is 1.7K lines long) and also, for a independent server and client implementation of mavlink_parameters (which would also increase readability) this will be re-usable.
