+++ src/mavsdk/core/mavlink_parameters.cpp
@@ -7,6 +7,7 @@
 #include <future>
 
 namespace mavsdk {
+using namespace parameters;
 
 MAVLinkParameters::MAVLinkParameters(
     Sender& sender,
@@ -67,6 +68,10 @@ MAVLinkParameters::MAVLinkParameters(
             MAVLINK_MSG_ID_PARAM_REQUEST_LIST,
             [this](const mavlink_message_t& message) { process_param_request_list(message); },
             this);
+		_message_handler.register_one(
+			MAVLINK_MSG_ID_PARAM_EXT_REQUEST_LIST,
+			[this](const mavlink_message_t& message) { process_param_ext_request_list(message); },
+			this);
 
         _message_handler.register_one(
             MAVLINK_MSG_ID_PARAM_EXT_REQUEST_READ,
@@ -200,7 +205,7 @@ void MAVLinkParameters::set_param_int_async(
     const auto set_step = [=]() {
         auto new_work = std::make_shared<WorkItem>(_timeout_s_callback());
 
-        MAVLinkParameters::ParamValue value_to_set;
+	  	parameters::ParamValue value_to_set;
         value_to_set.set(value);
 
         new_work->type = WorkItem::Type::Set;
@@ -262,7 +267,7 @@ void MAVLinkParameters::set_param_float_async(
 
     auto new_work = std::make_shared<WorkItem>(_timeout_s_callback());
 
-    MAVLinkParameters::ParamValue value_to_set;
+  	parameters::ParamValue value_to_set;
     value_to_set.set_float(value);
 
     new_work->type = WorkItem::Type::Set;
@@ -457,7 +462,7 @@ void MAVLinkParameters::set_param_custom_async(
 
     auto new_work = std::make_shared<WorkItem>(_timeout_s_callback());
 
-    MAVLinkParameters::ParamValue value_to_set;
+  	parameters::ParamValue value_to_set;
     value_to_set.set_custom(value);
 
     new_work->type = WorkItem::Type::Set;
@@ -470,13 +475,13 @@ void MAVLinkParameters::set_param_custom_async(
     _work_queue.push_back(new_work);
 }
 
-std::map<std::string, MAVLinkParameters::ParamValue> MAVLinkParameters::retrieve_all_server_params()
+std::map<std::string, parameters::ParamValue> MAVLinkParameters::retrieve_all_server_params()
 {
     std::lock_guard<std::mutex> lock(_all_params_mutex);
     return _all_params;
 }
 
-std::pair<MAVLinkParameters::Result, MAVLinkParameters::ParamValue>
+std::pair<MAVLinkParameters::Result, parameters::ParamValue>
 MAVLinkParameters::retrieve_server_param(const std::string& name, ParamValue value_type)
 {
     if (_all_params.find(name) != _all_params.end()) {
@@ -528,8 +533,8 @@ MAVLinkParameters::retrieve_server_param_int(const std::string& name)
     return {MAVLinkParameters::Result::NotFound, {}};
 }
 
-std::pair<MAVLinkParameters::Result, MAVLinkParameters::ParamValue> MAVLinkParameters::get_param(
-    const std::string& name, MAVLinkParameters::ParamValue value, bool extended)
+std::pair<MAVLinkParameters::Result, parameters::ParamValue> MAVLinkParameters::get_param(
+    const std::string& name, parameters::ParamValue value, bool extended)
 {
     auto prom = std::promise<std::pair<Result, ParamValue>>();
     auto res = prom.get_future();
@@ -594,31 +599,38 @@ MAVLinkParameters::get_param_custom(const std::string& name)
     return res.get();
 }
 
-void MAVLinkParameters::get_all_params_async(const GetAllParamsCallback& callback)
+void MAVLinkParameters::get_all_params_async(const GetAllParamsCallback& callback,const bool use_extended)
 {
     std::lock_guard<std::mutex> lock(_all_params_mutex);
-
     _all_params_callback = callback;
-
     mavlink_message_t msg;
-
-    mavlink_msg_param_request_list_pack(
-        _sender.get_own_system_id(),
-        _sender.get_own_component_id(),
-        &msg,
-        _sender.get_system_id(),
-        MAV_COMP_ID_AUTOPILOT1); // FIXME: what should the component be?
-
+	if(use_extended){
+	  mavlink_msg_param_request_list_pack(
+		  _sender.get_own_system_id(),
+		  _sender.get_own_component_id(),
+		  &msg,
+		  _sender.get_system_id(),
+		  MAV_COMP_ID_AUTOPILOT1); // FIXME: what should the component be?
+	}else{
+	  mavlink_msg_param_ext_request_list_pack(
+		  _sender.get_own_system_id(),
+		  _sender.get_own_component_id(),
+		  &msg,
+		  _sender.get_system_id(),
+		  MAV_COMP_ID_AUTOPILOT1); // FIXME: what should the component be?
+	}
     if (!_sender.send_message(msg)) {
         LogErr() << "Failed to send param list request!";
         callback(std::map<std::string, ParamValue>{});
     }
-
+	// There are 2 possible cases - we get all the messages in time - in this case, the
+	// _all_params_callback member is called. Otherwise, we get a timeout at some point, which then calls
+	// the _all_params_callback member with an empty result.
     _timeout_handler.add(
         [this] { receive_timeout(); }, _timeout_s_callback(), &_all_params_timeout_cookie);
 }
 
-std::map<std::string, MAVLinkParameters::ParamValue> MAVLinkParameters::get_all_params()
+std::map<std::string, parameters::ParamValue> MAVLinkParameters::get_all_params(const bool use_extended)
 {
     std::promise<std::map<std::string, ParamValue>> prom;
     auto res = prom.get_future();
@@ -628,9 +640,9 @@ std::map<std::string, MAVLinkParameters::ParamValue> MAVLinkParameters::get_all_
 		// Since for example on a receive timeout, the empty all_params result is constructed in-place and then
 		// goes out of scope when the callback returns. So don't use a reference here, pass by value or refactor the code to use
 		// a std::shared_ptr.
-        [&prom](std::map<std::string, MAVLinkParameters::ParamValue> all_params) {
+        [&prom](std::map<std::string, parameters::ParamValue> all_params) {
             prom.set_value(all_params);
-        });
+        },use_extended);
     return res.get();
 }
 
@@ -988,7 +1000,7 @@ void MAVLinkParameters::process_param_value(const mavlink_message_t& message)
         received_value.set_from_mavlink_param_value_bytewise(param_value);
     }
 
-    std::string param_id = extract_safe_param_id(param_value.param_id);
+    const std::string param_id = extract_safe_param_id(param_value.param_id);
 
     {
         std::lock_guard<std::mutex> lock(_all_params_mutex);
@@ -1318,7 +1330,7 @@ void MAVLinkParameters::process_param_ext_set(const mavlink_message_t& message)
         {
             // Use the ID
             std::lock_guard<std::mutex> lock(_all_params_mutex);
-            ParamValue value{};
+		  	parameters::ParamValue value{};
             if (!value.set_from_mavlink_param_ext_set(set_request)) {
                 LogWarn() << "Invalid Param Ext Set Request: " << safe_param_id;
                 return;
@@ -1379,7 +1391,7 @@ void MAVLinkParameters::receive_timeout()
 
     switch (work->type) {
         case WorkItem::Type::Get: {
-            ParamValue empty_value;
+		  	parameters::ParamValue empty_value;
             if (work->retries_to_do > 0) {
                 // We're not sure the command arrived, let's retransmit.
                 LogWarn() << "sending again, retries to do: " << work->retries_to_do << "  ("
@@ -1401,7 +1413,7 @@ void MAVLinkParameters::receive_timeout()
                     } else if (std::get_if<GetParamAnyCallback>(&work->callback)) {
                         const auto& callback = std::get<GetParamAnyCallback>(work->callback);
                         if (callback) {
-                            callback(Result::ConnectionError, ParamValue{});
+                            callback(Result::ConnectionError, parameters::ParamValue{});
                         }
                     }
                 } else {
@@ -1474,12 +1486,6 @@ std::string MAVLinkParameters::extract_safe_param_id(const char param_id[])
     return {param_id_long_enough};
 }
 
-std::ostream& operator<<(std::ostream& strm, const MAVLinkParameters::ParamValue& obj)
-{
-    strm << obj.get_string();
-    return strm;
-}
-
 void MAVLinkParameters::process_param_set(const mavlink_message_t& message)
 {
     mavlink_param_set_t set_request{};
@@ -1492,7 +1498,7 @@ void MAVLinkParameters::process_param_set(const mavlink_message_t& message)
 
         // Use the ID
         if (_all_params.find(safe_param_id) != _all_params.end()) {
-            ParamValue value{};
+		  	parameters::ParamValue value{};
             if (!value.set_from_mavlink_param_set_bytewise(set_request)) {
                 LogWarn() << "Invalid Param Set Request: " << safe_param_id;
                 return;
@@ -1577,8 +1583,7 @@ void MAVLinkParameters::process_param_request_list(const mavlink_message_t& mess
 {
     mavlink_param_request_list_t list_request{};
     mavlink_msg_param_request_list_decode(&message, &list_request);
-
-    auto idx = 0;
+	int idx=0;
     for (const auto& pair : _all_params) {
 	  	// make sure extended parameters never make it out via the param request list
 		// (use param extended list)
@@ -1598,6 +1603,23 @@ void MAVLinkParameters::process_param_request_list(const mavlink_message_t& mess
         new_work->param_index = idx++;
         _work_queue.push_back(new_work);
     }
+}
+
+void MAVLinkParameters::process_param_ext_request_list(const mavlink_message_t& message)
+{
+  mavlink_param_ext_request_list_t ext_list_request{};
+  mavlink_msg_param_ext_request_list_decode(&message, &ext_list_request);
+  int idx=0;
+  for (const auto& pair : _all_params) {
+	auto new_work = std::make_shared<WorkItem>(_timeout_s_callback());
+	new_work->type = WorkItem::Type::Value;
+	new_work->param_name = pair.first;
+	new_work->param_value = pair.second;
+	new_work->extended = true;
+	new_work->param_count = get_current_parameters_count(true);
+	new_work->param_index = idx++;
+	_work_queue.push_back(new_work);
+  }
 }
 
 void MAVLinkParameters::process_param_ext_request_read(const mavlink_message_t& message)
@@ -1624,624 +1646,7 @@ void MAVLinkParameters::process_param_ext_request_read(const mavlink_message_t& 
     }
 }
 
-bool MAVLinkParameters::ParamValue::set_from_mavlink_param_value_bytewise(
-    const mavlink_param_value_t& mavlink_value)
-{
-    switch (mavlink_value.param_type) {
-        case MAV_PARAM_TYPE_UINT8: {
-            uint8_t temp;
-            memcpy(&temp, &mavlink_value.param_value, sizeof(temp));
-            _value = temp;
-        } break;
 
-        case MAV_PARAM_TYPE_INT8: {
-            int8_t temp;
-            memcpy(&temp, &mavlink_value.param_value, sizeof(temp));
-            _value = temp;
-        } break;
-
-        case MAV_PARAM_TYPE_UINT16: {
-            uint16_t temp;
-            memcpy(&temp, &mavlink_value.param_value, sizeof(temp));
-            _value = temp;
-        } break;
-
-        case MAV_PARAM_TYPE_INT16: {
-            int16_t temp;
-            memcpy(&temp, &mavlink_value.param_value, sizeof(temp));
-            _value = temp;
-        } break;
-
-        case MAV_PARAM_TYPE_UINT32: {
-            uint32_t temp;
-            memcpy(&temp, &mavlink_value.param_value, sizeof(temp));
-            _value = temp;
-        } break;
-
-        case MAV_PARAM_TYPE_INT32: {
-            int32_t temp;
-            memcpy(&temp, &mavlink_value.param_value, sizeof(temp));
-            _value = temp;
-        } break;
-
-        case MAV_PARAM_TYPE_REAL32: {
-            _value = mavlink_value.param_value;
-        } break;
-
-        default:
-            // This would be worrying
-            LogErr() << "Error: unknown mavlink param type: "
-                     << std::to_string(mavlink_value.param_type);
-            return false;
-    }
-    return true;
-}
-
-bool MAVLinkParameters::ParamValue::set_from_mavlink_param_value_cast(
-    const mavlink_param_value_t& mavlink_value)
-{
-    switch (mavlink_value.param_type) {
-        case MAV_PARAM_TYPE_UINT8: {
-            _value = static_cast<uint8_t>(mavlink_value.param_value);
-        } break;
-
-        case MAV_PARAM_TYPE_INT8: {
-            _value = static_cast<int8_t>(mavlink_value.param_value);
-        } break;
-
-        case MAV_PARAM_TYPE_UINT16: {
-            _value = static_cast<uint16_t>(mavlink_value.param_value);
-        } break;
-
-        case MAV_PARAM_TYPE_INT16: {
-            _value = static_cast<int16_t>(mavlink_value.param_value);
-        } break;
-
-        case MAV_PARAM_TYPE_UINT32: {
-            _value = static_cast<uint32_t>(mavlink_value.param_value);
-        } break;
-
-        case MAV_PARAM_TYPE_INT32: {
-            _value = static_cast<int32_t>(mavlink_value.param_value);
-        } break;
-
-        case MAV_PARAM_TYPE_REAL32: {
-            float temp = mavlink_value.param_value;
-            _value = temp;
-        } break;
-
-        default:
-            // This would be worrying
-            LogErr() << "Error: unknown mavlink param type: "
-                     << std::to_string(mavlink_value.param_type);
-            return false;
-    }
-    return true;
-}
-
-bool MAVLinkParameters::ParamValue::set_from_mavlink_param_set_bytewise(
-    const mavlink_param_set_t& mavlink_set)
-{
-    mavlink_param_value_t mavlink_value{};
-    mavlink_value.param_value = mavlink_set.param_value;
-    mavlink_value.param_type = mavlink_set.param_type;
-
-    return set_from_mavlink_param_value_bytewise(mavlink_value);
-}
-
-bool MAVLinkParameters::ParamValue::set_from_mavlink_param_ext_set(
-    const mavlink_param_ext_set_t& mavlink_ext_set)
-{
-    switch (mavlink_ext_set.param_type) {
-        case MAV_PARAM_EXT_TYPE_UINT8: {
-            uint8_t temp;
-            memcpy(&temp, &mavlink_ext_set.param_value[0], sizeof(temp));
-            _value = temp;
-        } break;
-        case MAV_PARAM_EXT_TYPE_INT8: {
-            int8_t temp;
-            memcpy(&temp, &mavlink_ext_set.param_value[0], sizeof(temp));
-            _value = temp;
-        } break;
-        case MAV_PARAM_EXT_TYPE_UINT16: {
-            uint16_t temp;
-            memcpy(&temp, &mavlink_ext_set.param_value[0], sizeof(temp));
-            _value = temp;
-        } break;
-        case MAV_PARAM_EXT_TYPE_INT16: {
-            int16_t temp;
-            memcpy(&temp, &mavlink_ext_set.param_value[0], sizeof(temp));
-            _value = temp;
-        } break;
-        case MAV_PARAM_EXT_TYPE_UINT32: {
-            uint32_t temp;
-            memcpy(&temp, &mavlink_ext_set.param_value[0], sizeof(temp));
-            _value = temp;
-        } break;
-        case MAV_PARAM_EXT_TYPE_INT32: {
-            int32_t temp;
-            memcpy(&temp, &mavlink_ext_set.param_value[0], sizeof(temp));
-            _value = temp;
-        } break;
-        case MAV_PARAM_EXT_TYPE_UINT64: {
-            uint64_t temp;
-            memcpy(&temp, &mavlink_ext_set.param_value[0], sizeof(temp));
-            _value = temp;
-        } break;
-        case MAV_PARAM_EXT_TYPE_INT64: {
-            int64_t temp;
-            memcpy(&temp, &mavlink_ext_set.param_value[0], sizeof(temp));
-            _value = temp;
-        } break;
-        case MAV_PARAM_EXT_TYPE_REAL32: {
-            float temp;
-            memcpy(&temp, &mavlink_ext_set.param_value[0], sizeof(temp));
-            _value = temp;
-        } break;
-        case MAV_PARAM_EXT_TYPE_REAL64: {
-            double temp;
-            memcpy(&temp, &mavlink_ext_set.param_value[0], sizeof(temp));
-            _value = temp;
-        } break;
-        case MAV_PARAM_EXT_TYPE_CUSTOM: {
-            std::size_t len = std::min(std::size_t(128), strlen(mavlink_ext_set.param_value));
-            _value = std::string(mavlink_ext_set.param_value, mavlink_ext_set.param_value + len);
-        } break;
-        default:
-            // This would be worrying
-            LogErr() << "Error: unknown mavlink ext param type";
-            assert(false);
-            return false;
-    }
-    return true;
-}
-
-// FIXME: this function can collapse with the one above.
-bool MAVLinkParameters::ParamValue::set_from_mavlink_param_ext_value(
-    const mavlink_param_ext_value_t& mavlink_ext_value)
-{
-    switch (mavlink_ext_value.param_type) {
-        case MAV_PARAM_EXT_TYPE_UINT8: {
-            uint8_t temp;
-            memcpy(&temp, &mavlink_ext_value.param_value[0], sizeof(temp));
-            _value = temp;
-        } break;
-        case MAV_PARAM_EXT_TYPE_INT8: {
-            int8_t temp;
-            memcpy(&temp, &mavlink_ext_value.param_value[0], sizeof(temp));
-            _value = temp;
-        } break;
-        case MAV_PARAM_EXT_TYPE_UINT16: {
-            uint16_t temp;
-            memcpy(&temp, &mavlink_ext_value.param_value[0], sizeof(temp));
-            _value = temp;
-        } break;
-        case MAV_PARAM_EXT_TYPE_INT16: {
-            int16_t temp;
-            memcpy(&temp, &mavlink_ext_value.param_value[0], sizeof(temp));
-            _value = temp;
-        } break;
-        case MAV_PARAM_EXT_TYPE_UINT32: {
-            uint32_t temp;
-            memcpy(&temp, &mavlink_ext_value.param_value[0], sizeof(temp));
-            _value = temp;
-        } break;
-        case MAV_PARAM_EXT_TYPE_INT32: {
-            int32_t temp;
-            memcpy(&temp, &mavlink_ext_value.param_value[0], sizeof(temp));
-            _value = temp;
-        } break;
-        case MAV_PARAM_EXT_TYPE_UINT64: {
-            uint64_t temp;
-            memcpy(&temp, &mavlink_ext_value.param_value[0], sizeof(temp));
-            _value = temp;
-        } break;
-        case MAV_PARAM_EXT_TYPE_INT64: {
-            int64_t temp;
-            memcpy(&temp, &mavlink_ext_value.param_value[0], sizeof(temp));
-            _value = temp;
-        } break;
-        case MAV_PARAM_EXT_TYPE_REAL32: {
-            float temp;
-            memcpy(&temp, &mavlink_ext_value.param_value[0], sizeof(temp));
-            _value = temp;
-        } break;
-        case MAV_PARAM_EXT_TYPE_REAL64: {
-            double temp;
-            memcpy(&temp, &mavlink_ext_value.param_value[0], sizeof(temp));
-            _value = temp;
-        } break;
-        case MAV_PARAM_EXT_TYPE_CUSTOM: {
-            std::size_t len = std::min(std::size_t(128), strlen(mavlink_ext_value.param_value));
-            _value =
-                std::string(mavlink_ext_value.param_value, mavlink_ext_value.param_value + len);
-        } break;
-        default:
-            // This would be worrying
-            LogErr() << "Error: unknown mavlink ext param type";
-            assert(false);
-            return false;
-    }
-    return true;
-}
-
-bool MAVLinkParameters::ParamValue::set_from_xml(
-    const std::string& type_str, const std::string& value_str)
-{
-    if (strcmp(type_str.c_str(), "uint8") == 0) {
-        _value = static_cast<uint8_t>(std::stoi(value_str));
-    } else if (strcmp(type_str.c_str(), "int8") == 0) {
-        _value = static_cast<int8_t>(std::stoi(value_str));
-    } else if (strcmp(type_str.c_str(), "uint16") == 0) {
-        _value = static_cast<uint16_t>(std::stoi(value_str));
-    } else if (strcmp(type_str.c_str(), "int16") == 0) {
-        _value = static_cast<int16_t>(std::stoi(value_str));
-    } else if (strcmp(type_str.c_str(), "uint32") == 0) {
-        _value = static_cast<uint32_t>(std::stoi(value_str));
-    } else if (strcmp(type_str.c_str(), "int32") == 0) {
-        _value = static_cast<int32_t>(std::stoi(value_str));
-    } else if (strcmp(type_str.c_str(), "uint64") == 0) {
-        _value = static_cast<uint64_t>(std::stoll(value_str));
-    } else if (strcmp(type_str.c_str(), "int64") == 0) {
-        _value = static_cast<int64_t>(std::stoll(value_str));
-    } else if (strcmp(type_str.c_str(), "float") == 0) {
-        _value = static_cast<float>(std::stof(value_str));
-    } else if (strcmp(type_str.c_str(), "double") == 0) {
-        _value = static_cast<double>(std::stod(value_str));
-    } else {
-        LogErr() << "Unknown type: " << type_str;
-        return false;
-    }
-    return true;
-}
-
-bool MAVLinkParameters::ParamValue::set_empty_type_from_xml(const std::string& type_str)
-{
-    if (strcmp(type_str.c_str(), "uint8") == 0) {
-        _value = uint8_t(0);
-    } else if (strcmp(type_str.c_str(), "int8") == 0) {
-        _value = int8_t(0);
-    } else if (strcmp(type_str.c_str(), "uint16") == 0) {
-        _value = uint16_t(0);
-    } else if (strcmp(type_str.c_str(), "int16") == 0) {
-        _value = int16_t(0);
-    } else if (strcmp(type_str.c_str(), "uint32") == 0) {
-        _value = uint32_t(0);
-    } else if (strcmp(type_str.c_str(), "int32") == 0) {
-        _value = int32_t(0);
-    } else if (strcmp(type_str.c_str(), "uint64") == 0) {
-        _value = uint64_t(0);
-    } else if (strcmp(type_str.c_str(), "int64") == 0) {
-        _value = int64_t(0);
-    } else if (strcmp(type_str.c_str(), "float") == 0) {
-        _value = 0.0f;
-    } else if (strcmp(type_str.c_str(), "double") == 0) {
-        _value = 0.0;
-    } else {
-        LogErr() << "Unknown type: " << type_str;
-        return false;
-    }
-    return true;
-}
-
-[[nodiscard]] MAV_PARAM_TYPE MAVLinkParameters::ParamValue::get_mav_param_type() const
-{
-    if (std::get_if<uint8_t>(&_value)) {
-        return MAV_PARAM_TYPE_UINT8;
-    } else if (std::get_if<int8_t>(&_value)) {
-        return MAV_PARAM_TYPE_INT8;
-    } else if (std::get_if<uint16_t>(&_value)) {
-        return MAV_PARAM_TYPE_UINT16;
-    } else if (std::get_if<int16_t>(&_value)) {
-        return MAV_PARAM_TYPE_INT16;
-    } else if (std::get_if<uint32_t>(&_value)) {
-        return MAV_PARAM_TYPE_UINT32;
-    } else if (std::get_if<int32_t>(&_value)) {
-        return MAV_PARAM_TYPE_INT32;
-    } else if (std::get_if<uint64_t>(&_value)) {
-        return MAV_PARAM_TYPE_UINT64;
-    } else if (std::get_if<int64_t>(&_value)) {
-        return MAV_PARAM_TYPE_INT64;
-    } else if (std::get_if<float>(&_value)) {
-        return MAV_PARAM_TYPE_REAL32;
-    } else if (std::get_if<double>(&_value)) {
-        return MAV_PARAM_TYPE_REAL64;
-    } else {
-        LogErr() << "Unknown data type for param.";
-        assert(false);
-        return MAV_PARAM_TYPE_INT32;
-    }
-}
-
-[[nodiscard]] MAV_PARAM_EXT_TYPE MAVLinkParameters::ParamValue::get_mav_param_ext_type() const
-{
-    if (std::get_if<uint8_t>(&_value)) {
-        return MAV_PARAM_EXT_TYPE_UINT8;
-    } else if (std::get_if<int8_t>(&_value)) {
-        return MAV_PARAM_EXT_TYPE_INT8;
-    } else if (std::get_if<uint16_t>(&_value)) {
-        return MAV_PARAM_EXT_TYPE_UINT16;
-    } else if (std::get_if<int16_t>(&_value)) {
-        return MAV_PARAM_EXT_TYPE_INT16;
-    } else if (std::get_if<uint32_t>(&_value)) {
-        return MAV_PARAM_EXT_TYPE_UINT32;
-    } else if (std::get_if<int32_t>(&_value)) {
-        return MAV_PARAM_EXT_TYPE_INT32;
-    } else if (std::get_if<uint64_t>(&_value)) {
-        return MAV_PARAM_EXT_TYPE_UINT64;
-    } else if (std::get_if<int64_t>(&_value)) {
-        return MAV_PARAM_EXT_TYPE_INT64;
-    } else if (std::get_if<float>(&_value)) {
-        return MAV_PARAM_EXT_TYPE_REAL32;
-    } else if (std::get_if<double>(&_value)) {
-        return MAV_PARAM_EXT_TYPE_REAL64;
-    } else if (std::get_if<std::string>(&_value)) {
-        return MAV_PARAM_EXT_TYPE_CUSTOM;
-    } else {
-        LogErr() << "Unknown data type for param.";
-        assert(false);
-        return MAV_PARAM_EXT_TYPE_INT32;
-    }
-}
-
-bool MAVLinkParameters::ParamValue::set_as_same_type(const std::string& value_str)
-{
-    if (std::get_if<uint8_t>(&_value)) {
-        _value = uint8_t(std::stoi(value_str));
-    } else if (std::get_if<int8_t>(&_value)) {
-        _value = int8_t(std::stoi(value_str));
-    } else if (std::get_if<uint16_t>(&_value)) {
-        _value = uint16_t(std::stoi(value_str));
-    } else if (std::get_if<int16_t>(&_value)) {
-        _value = int16_t(std::stoi(value_str));
-    } else if (std::get_if<uint32_t>(&_value)) {
-        _value = uint32_t(std::stoi(value_str));
-    } else if (std::get_if<int32_t>(&_value)) {
-        _value = int32_t(std::stoi(value_str));
-    } else if (std::get_if<uint64_t>(&_value)) {
-        _value = uint64_t(std::stoll(value_str));
-    } else if (std::get_if<int64_t>(&_value)) {
-        _value = int64_t(std::stoll(value_str));
-    } else if (std::get_if<float>(&_value)) {
-        _value = float(std::stof(value_str));
-    } else if (std::get_if<double>(&_value)) {
-        _value = double(std::stod(value_str));
-    } else {
-        LogErr() << "Unknown type";
-        return false;
-    }
-    return true;
-}
-
-[[nodiscard]] float MAVLinkParameters::ParamValue::get_4_float_bytes_bytewise() const
-{
-    if (std::get_if<float>(&_value)) {
-        return std::get<float>(_value);
-    } else if (std::get_if<int32_t>(&_value)) {
-        return *(reinterpret_cast<const float*>(&std::get<int32_t>(_value)));
-    } else {
-        LogErr() << "Unknown type";
-        assert(false);
-        return NAN;
-    }
-}
-
-[[nodiscard]] float MAVLinkParameters::ParamValue::get_4_float_bytes_cast() const
-{
-    if (std::get_if<float>(&_value)) {
-        return std::get<float>(_value);
-    } else if (std::get_if<uint32_t>(&_value)) {
-        return static_cast<float>(std::get<uint32_t>(_value));
-    } else if (std::get_if<int32_t>(&_value)) {
-        return static_cast<float>(std::get<int32_t>(_value));
-    } else if (std::get_if<uint16_t>(&_value)) {
-        return static_cast<float>(std::get<uint16_t>(_value));
-    } else if (std::get_if<int16_t>(&_value)) {
-        return static_cast<float>(std::get<int16_t>(_value));
-    } else if (std::get_if<uint8_t>(&_value)) {
-        return static_cast<float>(std::get<uint8_t>(_value));
-    } else if (std::get_if<int8_t>(&_value)) {
-        return static_cast<float>(std::get<int8_t>(_value));
-    } else {
-        LogErr() << "Unknown type";
-        assert(false);
-        return NAN;
-    }
-}
-
-[[nodiscard]] std::optional<int> MAVLinkParameters::ParamValue::get_int() const
-{
-    if (std::get_if<uint32_t>(&_value)) {
-        return static_cast<int>(std::get<uint32_t>(_value));
-    } else if (std::get_if<int32_t>(&_value)) {
-        return static_cast<int>(std::get<int32_t>(_value));
-    } else if (std::get_if<uint16_t>(&_value)) {
-        return static_cast<int>(std::get<uint16_t>(_value));
-    } else if (std::get_if<int16_t>(&_value)) {
-        return static_cast<int>(std::get<int16_t>(_value));
-    } else if (std::get_if<uint8_t>(&_value)) {
-        return static_cast<int>(std::get<uint8_t>(_value));
-    } else if (std::get_if<int8_t>(&_value)) {
-        return static_cast<int>(std::get<int8_t>(_value));
-    } else {
-        LogErr() << "Not int type";
-        return {};
-    }
-}
-
-bool MAVLinkParameters::ParamValue::set_int(int new_value)
-{
-    if (std::get_if<uint8_t>(&_value)) {
-        _value = static_cast<uint8_t>(new_value);
-        return true;
-    } else if (std::get_if<int8_t>(&_value)) {
-        _value = static_cast<int8_t>(new_value);
-        return true;
-    } else if (std::get_if<uint16_t>(&_value)) {
-        _value = static_cast<uint16_t>(new_value);
-        return true;
-    } else if (std::get_if<int16_t>(&_value)) {
-        _value = static_cast<int16_t>(new_value);
-        return true;
-    } else if (std::get_if<uint32_t>(&_value)) {
-        _value = static_cast<uint32_t>(new_value);
-        return true;
-    } else if (std::get_if<int32_t>(&_value)) {
-        _value = static_cast<int32_t>(new_value);
-        return true;
-    } else {
-        return false;
-    }
-}
-
-void MAVLinkParameters::ParamValue::set_float(float new_value)
-{
-    _value = new_value;
-}
-
-void MAVLinkParameters::ParamValue::set_custom(const std::string& new_value)
-{
-    _value = new_value;
-}
-
-[[nodiscard]] std::optional<float> MAVLinkParameters::ParamValue::get_float() const
-{
-    if (std::get_if<float>(&_value)) {
-        return std::get<float>(_value);
-    } else {
-        LogErr() << "Not float type";
-        return {};
-    }
-}
-
-[[nodiscard]] std::optional<std::string> MAVLinkParameters::ParamValue::get_custom() const
-{
-    if (std::get_if<std::string>(&_value)) {
-        return std::get<std::string>(_value);
-    } else {
-        LogErr() << "Not custom type";
-        return {};
-    }
-}
-
-std::array<char, 128> MAVLinkParameters::ParamValue::get_128_bytes() const
-{
-    std::array<char, 128> bytes{};
-
-    if (std::get_if<uint8_t>(&_value)) {
-        memcpy(bytes.data(), &std::get<uint8_t>(_value), sizeof(uint8_t));
-    } else if (std::get_if<int8_t>(&_value)) {
-        memcpy(bytes.data(), &std::get<int8_t>(_value), sizeof(int8_t));
-    } else if (std::get_if<uint16_t>(&_value)) {
-        memcpy(bytes.data(), &std::get<uint16_t>(_value), sizeof(uint16_t));
-    } else if (std::get_if<int16_t>(&_value)) {
-        memcpy(bytes.data(), &std::get<int16_t>(_value), sizeof(int16_t));
-    } else if (std::get_if<uint32_t>(&_value)) {
-        memcpy(bytes.data(), &std::get<uint32_t>(_value), sizeof(uint32_t));
-    } else if (std::get_if<int32_t>(&_value)) {
-        memcpy(bytes.data(), &std::get<int32_t>(_value), sizeof(int32_t));
-    } else if (std::get_if<uint64_t>(&_value)) {
-        memcpy(bytes.data(), &std::get<uint64_t>(_value), sizeof(uint64_t));
-    } else if (std::get_if<int64_t>(&_value)) {
-        memcpy(bytes.data(), &std::get<int64_t>(_value), sizeof(int64_t));
-    } else if (std::get_if<float>(&_value)) {
-        memcpy(bytes.data(), &std::get<float>(_value), sizeof(float));
-    } else if (std::get_if<double>(&_value)) {
-        memcpy(bytes.data(), &std::get<double>(_value), sizeof(double));
-    } else if (std::get_if<std::string>(&_value)) {
-        auto str_ptr = &std::get<std::string>(_value);
-        // Copy all data in string, max 128 bytes
-        memcpy(bytes.data(), str_ptr->data(), std::min(bytes.size(), str_ptr->size()));
-    } else {
-        LogErr() << "Unknown type";
-        assert(false);
-    }
-
-    return bytes;
-}
-
-[[nodiscard]] std::string MAVLinkParameters::ParamValue::get_string() const
-{
-    return std::visit([](auto value) { return to_string(value); }, _value);
-}
-
-[[nodiscard]] bool MAVLinkParameters::ParamValue::is_same_type(const ParamValue& rhs) const
-{
-    if ((std::get_if<uint8_t>(&_value) && std::get_if<uint8_t>(&rhs._value)) ||
-        (std::get_if<int8_t>(&_value) && std::get_if<int8_t>(&rhs._value)) ||
-        (std::get_if<uint16_t>(&_value) && std::get_if<uint16_t>(&rhs._value)) ||
-        (std::get_if<int16_t>(&_value) && std::get_if<int16_t>(&rhs._value)) ||
-        (std::get_if<uint32_t>(&_value) && std::get_if<uint32_t>(&rhs._value)) ||
-        (std::get_if<int32_t>(&_value) && std::get_if<int32_t>(&rhs._value)) ||
-        (std::get_if<uint64_t>(&_value) && std::get_if<uint64_t>(&rhs._value)) ||
-        (std::get_if<int64_t>(&_value) && std::get_if<int64_t>(&rhs._value)) ||
-        (std::get_if<float>(&_value) && std::get_if<float>(&rhs._value)) ||
-        (std::get_if<double>(&_value) && std::get_if<double>(&rhs._value)) ||
-        (std::get_if<std::string>(&_value) && std::get_if<std::string>(&rhs._value))) {
-        return true;
-    } else {
-        LogWarn() << "Comparison type mismatch between " << typestr() << " and " << rhs.typestr();
-        return false;
-    }
-}
-
-bool MAVLinkParameters::ParamValue::operator==(const std::string& value_str) const
-{
-    // LogDebug() << "Compare " << value_str() << " and " << rhs.value_str();
-    if (std::get_if<uint8_t>(&_value)) {
-        return std::get<uint8_t>(_value) == std::stoi(value_str);
-    } else if (std::get_if<int8_t>(&_value)) {
-        return std::get<int8_t>(_value) == std::stoi(value_str);
-    } else if (std::get_if<uint16_t>(&_value)) {
-        return std::get<uint16_t>(_value) == std::stoi(value_str);
-    } else if (std::get_if<int16_t>(&_value)) {
-        return std::get<int16_t>(_value) == std::stoi(value_str);
-    } else if (std::get_if<uint32_t>(&_value)) {
-        return std::get<uint32_t>(_value) == std::stoul(value_str);
-    } else if (std::get_if<int32_t>(&_value)) {
-        return std::get<int32_t>(_value) == std::stol(value_str);
-    } else if (std::get_if<uint64_t>(&_value)) {
-        return std::get<uint64_t>(_value) == std::stoull(value_str);
-    } else if (std::get_if<int64_t>(&_value)) {
-        return std::get<int64_t>(_value) == std::stoll(value_str);
-    } else if (std::get_if<float>(&_value)) {
-        return std::get<float>(_value) == std::stof(value_str);
-    } else if (std::get_if<double>(&_value)) {
-        return std::get<double>(_value) == std::stod(value_str);
-    } else {
-        // This also covers custom_type_t
-        return false;
-    }
-}
-
-[[nodiscard]] std::string MAVLinkParameters::ParamValue::typestr() const
-{
-    if (std::get_if<uint8_t>(&_value)) {
-        return "uint8_t";
-    } else if (std::get_if<int8_t>(&_value)) {
-        return "int8_t";
-    } else if (std::get_if<uint16_t>(&_value)) {
-        return "uint16_t";
-    } else if (std::get_if<int16_t>(&_value)) {
-        return "int16_t";
-    } else if (std::get_if<uint32_t>(&_value)) {
-        return "uint32_t";
-    } else if (std::get_if<int32_t>(&_value)) {
-        return "int32_t";
-    } else if (std::get_if<uint64_t>(&_value)) {
-        return "uint64_t";
-    } else if (std::get_if<int64_t>(&_value)) {
-        return "int64_t";
-    } else if (std::get_if<float>(&_value)) {
-        return "float";
-    } else if (std::get_if<double>(&_value)) {
-        return "double";
-    } else if (std::get_if<std::string>(&_value)) {
-        return "custom";
-    }
-    // FIXME: Added to fix CI error (control reading end of non-void function)
-    return "unknown";
-}
 
 std::ostream& operator<<(std::ostream& str, const MAVLinkParameters::Result& result)
 {
