+++ src/mavsdk/plugins/camera/camera_impl.cpp
@@ -10,7 +10,7 @@
 #include <sstream>
 
 namespace mavsdk {
-
+using namespace parameters;
 CameraImpl::CameraImpl(System& system) : PluginImplBase(system)
 {
     _parent->register_plugin(this);
@@ -751,7 +751,7 @@ void CameraImpl::save_camera_mode(const float mavlink_camera_mode)
     // maintain (what if the MAVLink CAMERA_MODE enum evolves?), so
     // I am assuming here that in such a case, CAMERA_SETTINGS is
     // never sent by the camera.
-    MAVLinkParameters::ParamValue value;
+  parameters::ParamValue value;
     if (_camera_definition->get_setting("CAM_MODE", value)) {
         if (value.is<uint8_t>()) {
             value.set<uint8_t>(static_cast<uint8_t>(mavlink_camera_mode));
@@ -1395,7 +1395,7 @@ bool CameraImpl::get_possible_setting_options(std::vector<std::string>& settings
         return false;
     }
 
-    std::unordered_map<std::string, MAVLinkParameters::ParamValue> cd_settings{};
+    std::unordered_map<std::string, parameters::ParamValue> cd_settings{};
     _camera_definition->get_possible_settings(cd_settings);
 
     for (const auto& cd_setting : cd_settings) {
@@ -1420,7 +1420,7 @@ bool CameraImpl::get_possible_options(
         return false;
     }
 
-    std::vector<MAVLinkParameters::ParamValue> values;
+    std::vector<parameters::ParamValue> values;
     if (!_camera_definition->get_possible_options(setting_id, values)) {
         return false;
     }
@@ -1475,11 +1475,11 @@ void CameraImpl::set_option_async(
     }
 
     // We get it first so that we have the type of the param value.
-    MAVLinkParameters::ParamValue value;
+  parameters::ParamValue value;
 
     if (_camera_definition->is_setting_range(setting_id)) {
         // TODO: Get type from minimum.
-        std::vector<MAVLinkParameters::ParamValue> all_values;
+        std::vector<parameters::ParamValue> all_values;
         if (!_camera_definition->get_all_options(setting_id, all_values)) {
             if (callback) {
                 LogErr() << "Could not get all options to get type for range param.";
@@ -1523,7 +1523,7 @@ void CameraImpl::set_option_async(
             return;
         }
 
-        std::vector<MAVLinkParameters::ParamValue> possible_values;
+        std::vector<parameters::ParamValue> possible_values;
         _camera_definition->get_possible_options(setting_id, possible_values);
         bool allowed = false;
         for (const auto& possible_value : possible_values) {
@@ -1655,7 +1655,7 @@ void CameraImpl::get_option_async(
         return;
     }
 
-    MAVLinkParameters::ParamValue value;
+parameters::ParamValue value;
     // We should have this cached and don't need to get the param.
     if (_camera_definition->get_setting(setting_id, value)) {
         if (callback) {
@@ -1722,7 +1722,7 @@ void CameraImpl::notify_current_settings()
 
     for (auto& possible_setting : possible_setting_options) {
         // use the cache for this, presumably we updated it right before.
-        MAVLinkParameters::ParamValue value;
+	  parameters::ParamValue value;
         if (_camera_definition->get_setting(possible_setting, value)) {
             Camera::Setting setting{};
             setting.setting_id = possible_setting;
@@ -1794,7 +1794,7 @@ void CameraImpl::refresh_params()
         return;
     }
 
-    std::vector<std::pair<std::string, MAVLinkParameters::ParamValue>> params;
+    std::vector<std::pair<std::string, parameters::ParamValue>> params;
     _camera_definition->get_unknown_params(params);
     if (params.size() == 0) {
         // We're assuming that we changed one option and this did not cause
@@ -1808,13 +1808,13 @@ void CameraImpl::refresh_params()
     unsigned count = 0;
     for (const auto& param : params) {
         const std::string& param_name = param.first;
-        const MAVLinkParameters::ParamValue& param_value_type = param.second;
+        const parameters::ParamValue& param_value_type = param.second;
         const bool is_last = (count == params.size() - 1);
         _parent->get_param_async(
             param_name,
             param_value_type,
             [param_name, is_last, this](
-                MAVLinkParameters::Result result, MAVLinkParameters::ParamValue value) {
+                MAVLinkParameters::Result result, parameters::ParamValue value) {
                 if (result != MAVLinkParameters::Result::Success) {
                     return;
                 }
